Successful GETs for pics & names /handlebars templ
Button for Create Game generates working get requests for the pictures
table and the names table. The handlbars template for pictures renders
as expected BUT there is a problem getting the "names" data to generate
in the same template or element. Currently the names bank renders on top
of the pictures template as a display of a successful GET but it needs
to be with the form generated in flash-card.handlebars